fix(06-N6): address Rust 1.95 clippy lints surfaced by stable bump
Rust 1.95 (2026-04-14) introduced a number of new clippy lints that
surface across the workspace; the original 06-N6 chain landed before the
toolchain bump. This patch is the strict-CI follow-up:

Numerical / algorithm-faithful:
  - xcfun-kernels lib.rs: allow `excessive_precision` + `approx_constant`
    crate-wide (FreeBSD msun + xcfun-master constants must match the
    reference verbatim — algorithmic-identity rule, CLAUDE.md ACC).
  - xcfun-ad/src/expand/erf.rs: same allow at module level + replace the
    two `1.1283791670955126_f64` literals with std::f64::consts::FRAC_2_SQRT_PI
    (bit-equal value; reads as a named constant).
  - xcfun-ad/src/expand/br_inverse.rs: drop redundant `i as usize` casts
    (cubecl `for i in 2..size` already yields usize).

Mechanical clippy fixes:
  - br_like.rs: `x = x + step` -> `x += step` (assign_op).
  - revtpssx.rs / tpssx.rs / shared/tpss_like.rs: drop unnecessary
    `&mut` on `tpss_F_x` / `revtpss_fx` / `revtpss_H` arguments.
  - dispatch.rs: keep explicit `|`-joined ID list; allow `manual_range_patterns`
    so wave/phase grouping comments stay aligned.
  - optx/mod.rs: allow `module_inception` (mirrors xcfun-master file layout).
  - regen_registry.rs: collapse nested `if let`, swap `is_some+unwrap` for
    `if let`, drop `Ok(?)` pattern.
  - check_no_anyhow.rs / check_no_fma.rs: collapse nested `if`s into
    `&&` chains (Rust 2024 if-let chains).
  - self_tests.rs: rewrite `for k in ... { if ... return Some(k) } None`
    as `(0..=2).find(...)` (manual_find).
  - functional.rs: drop `(0 + off)` identity ops; switch the zeroing loop
    to `iter_mut().take(...)`.
  - backend.rs: allow `should_implement_trait` on `Backend::from_str`
    (no error type — `FromStr` impl would change the public surface).
  - batch_api_shape.rs: elide `'fun` lifetime on the assertion helper.

App-boundary crates (validation, xcfun-rs, xcfun-capi, examples):
  - Add file/crate-level `#![allow(...)]` for stylistic lints that fight
    against deliberately-explicit harness style: too_many_arguments,
    type_complexity, drop_non_drop, redundant_closure, vec_init_then_push,
    not_unsafe_ptr_arg_deref (FFI symbols), should_implement_trait,
    new_without_default, branches_sharing_code, manual_range_contains,
    if_same_then_else, identity_op/erasing_op (test pedagogy).

Doc-comment indentation (clippy::doc_lazy_continuation /
doc_overindented_list_items):
  - validation/build.rs, xtask/src/bin/{check_boundaries,regen_capi_header,
    regen_registry}.rs, crates/xcfun-kernels/src/functionals/{lda/pw92c,
    lda/pz81c,mgga/shared/{constants,tpss_like}}.rs, crates/xcfun-capi/src/types.rs,
    crates/xcfun-rs/tests/zero_alloc_strict.rs, crates/xcfun-eval/tests/self_tests.rs.

After: `cargo clippy --workspace --all-targets --locked -- -D warnings`
exits 0; `cargo fmt --all -- --check` exits 0; `cargo build --workspace
--locked` succeeds; library unit tests still pass.

Closes the strict-CI tail of plan 06-N6.

+++ crates/xcfun-ad/src/expand/br_inverse.rs
@@ -251,7 +251,7 @@ pub fn br_inverse_expand<F: Float>(t: &mut Array<F>, #[comptime] n: u32) {
             br_z_ctaylor::<F>(t, &mut f, n);
             // t[i] = -f[i] * t[1]  (== -f[i] * inv_f1; we re-read t[1] for
             //                       clarity / symmetry with the C++ source).
-            t[i as usize] = -f[i as usize] * t[1];
+            t[i] = -f[i] * t[1];
         }
     }
 }
